A lot of stuff has happened
I added a node server to possibly do something cool. It really should be
in a branch but oh well it's easy enough to remove. And the files can
still be run as usual from opening index.html in the public folder so...

Other than that a lot of progress has been made. Not sure what all
exactly but deffinitely the changing background photos, and the floating
content box for Employment.